# backend/docs-static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SendOTPResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendOTPRequest": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"paths": {
...
		"/api/auth/send-register-otp": {
			"post": {
				"operationId": "SendRegisterOTP",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or user already exists"
					}
				},
				"description": "Send OTP for user registration",
				"summary": "Send OTP to email for registration",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOTPRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/auth/send-login-otp": {
			"post": {
				"operationId": "SendLoginOTP",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or user not found"
					}
				},
				"description": "Send OTP for user login",
				"summary": "Send OTP to email for login",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOTPRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"info": {
		"title": "ParkFlow API",
		"version": "1.0.0",
		"description": "API for the ParkFlow application",
		"contact": {
			"name": "API Support",
			"email": "support@ParkFlow.com"
		}
	},
	"servers": [
		{
			"url": "/",
			"description": "Default API server"
		}
	]
}